requirements.txt
pytest-playwright
numpy